rename text in images (#273)
* Ultimate Multisite v.3

This is an update and collaboration to Ultimate Multisite, based on WP Ultimo, which was based on Pro Sites.

* Changed Base Images

Changed all base images to reflect the new Ultimate Multisite name.

* Added New Logo Version

Simply added a white version of the logo to display on dark backgrounds. This was done quickly and needs to be re-done later.

* Corrected Add-On Issues

Removed some embedded code for better and quicker integration. Also, preparing to add a few enhancements.

* update

* fix install instructions

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* update url

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* Fix switching templates after purchase

---------

Co-authored-by: BOOFOLA LLC <todds@boofola.xyz>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>